+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added the `Swap3` trait to provide the `swap3_bca` and `swap3_cab` functions
+  directly on a slice reference.
+
 ## [0.1.0] - 2023-07-10
 
 ### Internal